6665 Move StaticFooter up to /Components
