+++ src/components/About/index.ts
@@ -1,2 +1,1 @@
 export * from "./Feedback";
-export * from "./StaticPageFooter";
